Chore [#19] 목표 추가 셀 이름 변경
